Asssignment 5.2
Write a program that repeatedly reads numbers until the user enters “done”. Once “done” is entered, print out the maximum and minimum of the numbers. If the user enters anything other than a number, detect their mistake using try and except and print an error message and skip to the next number.